fix: define snapshot trait outside of arrow module
