+++ crates/deltalake-core/src/kernel/mod.rs
@@ -8,6 +8,7 @@ pub mod client;
 pub mod error;
 pub mod expressions;
 pub mod schema;
+pub mod snapshot;
 
 pub use actions::*;
 pub use error::*;
